requirements.txt
annotated-doc>=0.0.4
annotated-types>=0.7.0
anyio>=4.0.0
boto3>=1.34.0
click>=8.0.0
fastapi>=0.100.0
greenlet>=3.0.0
h11>=0.14.0
httpx>=0.27.0
idna>=3.0
psycopg[binary]>=3.1.18
playwright>=1.40.0
playwright-stealth>=1.0.0
pydantic>=2.0.0
pydantic_core>=2.0.0
pyee>=11.0.0
python-dotenv>=1.0.0
requests>=2.31.0
starlette>=0.27.0
typing-inspection>=0.4.2
typing_extensions>=4.0.0
uvicorn>=0.20.0
google-auth>=2.0.0
itsdangerous>=2.0.0
